First Commit
This is a complete interpreter that can now be further expanded upon. Currently it is capable of declaring a variable and storing only integers inside. It can declare and if statement and an else statement as long as you only compare whether two integers are equal to each other. It can declare print statements as long as the print statement is only printing an integer.